+++ module-frontend-changes/cart.php
@@ -5,6 +5,7 @@
 class Frontend_Cart {
 
   function __construct() {
+    // disable shipping calculator
     add_filter( 'woocommerce_cart_ready_to_calc_shipping', '__return_false', 99 );
     add_filter( 'woocommerce_coupons_enabled', '__return_false' );
 
@@ -17,6 +18,17 @@ class Frontend_Cart {
     // add checkout button
     remove_action( 'woocommerce_cart_collaterals', 'woocommerce_cart_totals', 10 );
     add_action( 'woocommerce_cart_actions', array($this, 'add_proceed_checkout_button'), 999 );
+
+    ///// ALERT BOX
+
+    add_filter( 'woocommerce_add_success', array($this, 'disable_alert_remove_cart') );
+
+    add_filter( 'woocommerce_add_to_cart_validation', array($this, 'add_close_button_in_alert') );
+    add_filter( 'wc_add_to_cart_message_html', array($this, 'add_close_button_in_alert') );
+    add_filter( 'woocommerce_add_error', array($this, 'add_close_button_in_alert') );
+
+    // change the button in alert to go straight to checkout
+    add_filter( 'wc_add_to_cart_message_html', array($this, 'added_to_cart_message'), null, 2 );
   }
 
   /*
@@ -55,5 +67,49 @@ class Frontend_Cart {
   */
   function add_proceed_checkout_button() {
     do_action( 'woocommerce_proceed_to_checkout' );
+  }
+
+  /*
+    Remove the alert when removing item from cart
+    @filter woocommerce_add_success
+
+    @param $message (str) - Default message
+    @return str - Modified message
+  */
+  function disable_alert_remove_cart( $message ) {
+    if( strpos( $message, 'Undo' ) ) {
+      return false;
+    }
+
+    return $message;
+  }
+
+  /*
+    Add X button to dismiss the message
+
+    @filter woocommerce_add_to_cart_validation
+    @filter wc_add_to_cart_message
+  */
+  function add_close_button_in_alert( $message ) {
+    $button = '<span class="woocommerce-message-close">×</span>';
+    return $button . $message;
+  }
+
+  /*
+    Change the alert message after adding to cart
+    @filter wc_add_to_cart_message
+
+    @param $message (str) - The default message
+    @param $product_id (int) - The product that's just added to cart
+    @return str - Modified message
+  */
+  function added_to_cart_message( $message, $product_id ) {
+    $real_message = preg_replace( '/<a\D+a>/', '', $message ); // without <a> tag
+    $button = sprintf(
+      '<a href="%s" class="button wc-forward">%s</a> ',
+      esc_url( wc_get_page_permalink('checkout') ), esc_html__( 'Continue Payment', 'my' )
+    );
+
+    return $button . $real_message;
   }
 }