+++ module-frontend-changes/myaccount.php
@@ -7,6 +7,9 @@ class Frontend_MyAccount {
 
     add_action( 'woocommerce_before_customer_login_form', array($this, 'add_form_wrapper') );
     add_action( 'woocommerce_after_customer_login_form', array($this, 'close_form_wrapper') );
+
+    add_action( 'h_after_login_form', array($this, 'add_toggle_register') );
+    add_action( 'h_before_register_form', array($this, 'add_toggle_login') );
   }
 
   /*
@@ -29,7 +32,7 @@ class Frontend_MyAccount {
     @action woocommerce_before_customer_login_form
   */
   function add_form_wrapper() {
-    echo '<div class="hoo-account-wrapper">';
+    echo '<div class="h-myaccount-wrapper">';
   }
 
   /*
@@ -37,5 +40,19 @@ class Frontend_MyAccount {
   */
   function close_form_wrapper() {
     echo '</div>';
+  }
+
+  /*
+    @action h_after_login_form
+  */
+  function add_toggle_register() {
+    echo '<p class="text-center"><a class="button-passive h-toggle-form">Register »</a></p>';
+  }
+
+  /*
+    @action h_before_register_form
+  */
+  function add_toggle_login() {
+    echo '<a class="button-passive h-toggle-form">« Login</a>';
   }
 }