+++ templates/myaccount/form-login.php
@@ -1,0 +1,109 @@
+<?php
+/*
+  Login Form
+
+  Changes:
+  1. Added `h_after_login_form` action to add "Register" button.
+  2. Added `h_before_register_form` action to add "Back to Login" button.
+	3. Added `register` ID to Registration form so it can be targetted when the URL contains #register.
+	4. Swapped the div position of Register and Login so we can hide Login form when the URL contains #register
+*/
+if ( ! defined( 'ABSPATH' ) ) {
+	exit; // Exit if accessed directly.
+}
+
+?>
+
+<?php wc_print_notices(); ?>
+
+<?php do_action( 'woocommerce_before_customer_login_form' ); ?>
+
+<?php if ( get_option( 'woocommerce_enable_myaccount_registration' ) === 'yes' ) : ?>
+
+<div class="u-columns col2-set" id="customer_login">
+
+	<div class="u-column2 col-2" id="register"> <?php #3 #4 ?>
+    <?php do_action( 'h_before_register_form' ); #2 ?>
+		<h2><?php esc_html_e( 'Register', 'woocommerce' ); ?></h2>
+
+		<form method="post" class="woocommerce-form woocommerce-form-register register">
+			<?php do_action( 'woocommerce_register_form_start' ); ?>
+
+			<?php if ( 'no' === get_option( 'woocommerce_registration_generate_username' ) ) : ?>
+
+				<p class="woocommerce-form-row woocommerce-form-row--wide form-row form-row-wide">
+					<label for="reg_username"><?php esc_html_e( 'Username', 'woocommerce' ); ?>&nbsp;<span class="required">*</span></label>
+					<input type="text" class="woocommerce-Input woocommerce-Input--text input-text" name="username" id="reg_username" autocomplete="username" value="<?php echo ( ! empty( $_POST['username'] ) ) ? esc_attr( wp_unslash( $_POST['username'] ) ) : ''; ?>" /><?php // @codingStandardsIgnoreLine ?>
+				</p>
+
+			<?php endif; ?>
+
+			<p class="woocommerce-form-row woocommerce-form-row--wide form-row form-row-wide">
+				<label for="reg_email"><?php esc_html_e( 'Email address', 'woocommerce' ); ?>&nbsp;<span class="required">*</span></label>
+				<input type="email" class="woocommerce-Input woocommerce-Input--text input-text" name="email" id="reg_email" autocomplete="email" value="<?php echo ( ! empty( $_POST['email'] ) ) ? esc_attr( wp_unslash( $_POST['email'] ) ) : ''; ?>" /><?php // @codingStandardsIgnoreLine ?>
+			</p>
+
+			<?php if ( 'no' === get_option( 'woocommerce_registration_generate_password' ) ) : ?>
+
+				<p class="woocommerce-form-row woocommerce-form-row--wide form-row form-row-wide">
+					<label for="reg_password"><?php esc_html_e( 'Password', 'woocommerce' ); ?>&nbsp;<span class="required">*</span></label>
+					<input type="password" class="woocommerce-Input woocommerce-Input--text input-text" name="password" id="reg_password" autocomplete="new-password" />
+				</p>
+
+			<?php endif; ?>
+
+			<?php do_action( 'woocommerce_register_form' ); ?>
+
+			<p class="woocommerce-FormRow form-row">
+				<?php wp_nonce_field( 'woocommerce-register', 'woocommerce-register-nonce' ); ?>
+				<button type="submit" class="woocommerce-Button button" name="register" value="<?php esc_attr_e( 'Register', 'woocommerce' ); ?>"><?php esc_html_e( 'Register', 'woocommerce' ); ?></button>
+			</p>
+
+			<?php do_action( 'woocommerce_register_form_end' ); ?>
+
+		</form>
+	</div>
+
+	<div class="u-column1 col-1">
+
+<?php endif; ?>
+
+	<h2><?php esc_html_e( 'Login', 'woocommerce' ); ?></h2>
+
+	<form class="woocommerce-form woocommerce-form-login login" method="post">
+
+		<?php do_action( 'woocommerce_login_form_start' ); ?>
+
+		<p class="woocommerce-form-row woocommerce-form-row--wide form-row form-row-wide">
+			<label for="username"><?php esc_html_e( 'Username or email address', 'woocommerce' ); ?>&nbsp;<span class="required">*</span></label>
+			<input type="text" class="woocommerce-Input woocommerce-Input--text input-text" name="username" id="username" autocomplete="username" value="<?php echo ( ! empty( $_POST['username'] ) ) ? esc_attr( wp_unslash( $_POST['username'] ) ) : ''; ?>" /><?php // @codingStandardsIgnoreLine ?>
+		</p>
+		<p class="woocommerce-form-row woocommerce-form-row--wide form-row form-row-wide">
+			<label for="password"><?php esc_html_e( 'Password', 'woocommerce' ); ?>&nbsp;<span class="required">*</span></label>
+			<input class="woocommerce-Input woocommerce-Input--text input-text" type="password" name="password" id="password" autocomplete="current-password" />
+		</p>
+
+		<?php do_action( 'woocommerce_login_form' ); ?>
+
+		<p class="form-row">
+			<?php wp_nonce_field( 'woocommerce-login', 'woocommerce-login-nonce' ); ?>
+			<button type="submit" class="woocommerce-Button button" name="login" value="<?php esc_attr_e( 'Log in', 'woocommerce' ); ?>"><?php esc_html_e( 'Log in', 'woocommerce' ); ?></button>
+			<label class="woocommerce-form__label woocommerce-form__label-for-checkbox inline">
+				<input class="woocommerce-form__input woocommerce-form__input-checkbox" name="rememberme" type="checkbox" id="rememberme" value="forever" /> <span><?php esc_html_e( 'Remember me', 'woocommerce' ); ?></span>
+			</label>
+		</p>
+		<p class="woocommerce-LostPassword lost_password">
+			<a href="<?php echo esc_url( wp_lostpassword_url() ); ?>"><?php esc_html_e( 'Lost your password?', 'woocommerce' ); ?></a>
+		</p>
+
+		<?php do_action( 'woocommerce_login_form_end' ); ?>
+
+	</form>
+
+  <?php do_action( 'h_after_login_form' ); #1 ?>
+
+<?php if ( get_option( 'woocommerce_enable_myaccount_registration' ) === 'yes' ) : ?>
+</div>
+<?php endif; ?>
+
+<?php do_action( 'woocommerce_after_customer_login_form' ); ?>